requirements.txt
asyncio==3.4.3
aiohttp==3.9.1
asyncpg==0.29.0
openai==1.12.0
beautifulsoup4==4.12.3
trafilatura==1.6.1
semantic-text-splitter==0.8.2
tiktoken==0.5.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
lxml==4.9.3
